+++ deployment/ccip/changeset/solana_v0_1_1/cs_ops_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && solana
+
 package solana_test
 
 import (
